:zap: refactoring enums
